Support more general file extensions
This is a bit of a rewrite to support file extensions other than '*.h5'.
Previously there were a list of hardcoded extensions to decide if the file was
HDF5. Now I just invoke h5ls and check the return code, if it's valid we
assume we're working with an HDF5 file.